+++ includes/test.hpp
@@ -1,4 +1,8 @@
 #ifndef TEST_HPP
 #define TEST_HPP
 
+void test_tasks();
+void test_proto();
+void test_cascade();
+
 #endif
